Fix reject command, add maintenance mode, code cleanup

+++ commands/approve.js
@@ -2,11 +2,19 @@ const Discord = require('discord.js');
 const Settings = require('../models/settings.js');
 const Suggestion = require('../models/suggestions.js');
 const { noSuggestionsPerms, noSuggestionsLogs, noPerms, maintenanceMode } = require('../utils/errors.js');
+const { owner } = require('../config.json');
 
 exports.run = async (client, message, args) => {
+
     const cmdName = client.commands.get('approve', 'help.name');
 
+    let perms = message.guild.me.permissions;
+    if (!perms.has('MANAGE_MESSAGES')) return message.channel.send('I can\'t delete messages! Make sure I have this permission: Manage Messages`').then(msg => msg.delete(5000));
+
     message.delete().catch(O_o=>{});
+
+    let status = cmdStatus.get('status');
+    if (status === 'off' && message.author.id !== owner)  return maintenanceMode(message.channel);
 
     Settings.findOne({
         guildID: message.guild.id,
@@ -21,15 +29,12 @@ exports.run = async (client, message, args) => {
 
         let admins = [];
         message.guild.members.forEach(collected => {
-            if (collected.hasPermission('MANAGE_GUILD') && !collected.user.bot) {
-                
-                admins.push(collected.id);
-            }
+            
+            if (collected.hasPermission('MANAGE_GUILD') && !collected.user.bot) return admins.push(collected.id);
+            
         });
 
-        if (!staffRoles) return message.channel.send('No staff roles exist! Please create them or contact a server administrator to handle suggestions.').then(message => {
-            message.delete(5000)
-        }).catch(err => console.log(err));
+        if (roles.length === 0 && !admins.includes(message.member.id)) return message.channel.send('No staff roles exist! Please create them or contact a server administrator to handle suggestions.').then(msg => msg.delete(5000));
         
         if (!admins.includes(message.member.id) && !message.member.roles.some(r => staffRoles.includes(r))) return noSuggestionsPerms(message.channel);
 
@@ -37,8 +42,8 @@ exports.run = async (client, message, args) => {
         const suggestionsLogs = message.guild.channels.find(c => c.name === res.suggestionsLogs) ||  message.guild.channels.find(c => c.toString() === res.suggestionsLogs);
         if (!suggestionsLogs) return noSuggestionsLogs(message.channel);
 
-        const id = args[0];
-        if (!id) return message.channel.send(`Usage: \`${res.prefix + cmdName} <id>\``).then(message => { message.delete(5000) }).catch(console.error);
+        let id = args[0];
+        if (!id) return message.channel.send(`Usage: \`${res.prefix + cmdName} <id>\``).then(msg => msg.delete(5000)).catch(console.error);
 
         Suggestion.findOne({
             guildID: message.guild.id,
@@ -90,12 +95,13 @@ exports.run = async (client, message, args) => {
                             .setColor('#00e640')
                             .setFooter(`sID: ${id}`)
                             .setTimestamp();
-        
-                        if (embed.footer.text.includes(id)) {
+                        
+                        let footer = embed.footer.text;
+                        if (footer.includes(id)) {
                             suggestionsChannel.fetchMessage(embed.message.id)
                                 .then(async m => {
                                     await m.edit(approvedEmbed).then(m.delete(5000));
-                                    await message.channel.send(`Suggestion **${args[0]}** has been approved.`).then(message => { message.delete(5000) }).catch(console.error);
+                                    await message.channel.send(`Suggestion **${args[0]}** has been approved.`).then(msg => msg.delete(5000)).catch(console.error);
                                     await suggestionsLogs.send(logsEmbed);
                                     await sUser.send(dmEmbed);
         
@@ -116,15 +122,15 @@ exports.run = async (client, message, args) => {
             });
         });
     });
-}
+};
 
 exports.conf = {
     aliases: [],
-    status: ''
-}
+    status: 'true'
+};
 
 exports.help = {
     name: "approve",
     description: "Approve a submitted suggestion via the suggestion ID",
     usage: "approve <ID>"
-}
+};
