+++ commands/maintenance.js
@@ -1,9 +1,11 @@
 const Discord = require('discord.js');
-const fs = require('fs');
 const { owner, prefix } = require('../config.json');
 const Settings = require('../models/settings');
 
 exports.run = async (client, message, args) => {
+
+    let perms = message.guild.me.permissions;
+    if (!perms.has('MANAGE_MESSAGES')) return message.channel.send('I can\'t delete messages! Make sure I have this permission: Manage Messages`').then(msg => msg.delete(5000));
 
     message.delete().catch(O_o=>{});
 
@@ -24,105 +26,90 @@ exports.run = async (client, message, args) => {
             return `${obj.username}#${obj.discriminator}`;
         }
 
-        let cmds = Array.from(client.commands.keys());
-        let aliases = Array.from(client.aliases.keys());
+        let status = cmdStatus.get('status');
         switch (args[0]) {
             case 'on':
-                message.channel.send('Please type `confirm` if you would like to proceed with beginning maintenance mode. You have 15 seconds.')
-                    .then(() => {
-                        message.delete().catch(O_o => {});
-                        message.channel.awaitMessages(res => res.content === 'confirm', {
+
+                if (status === 'on') return message.channel.send('Maintenance mode is already active.').then(msg => msg.delete(3003));
+
+                message.channel.send('Please type the maintenance message. You have 15 seconds.')
+                    .then(async () => {
+
+                        await message.delete().catch(O_o => {});
+
+                        message.channel.awaitMessages(res => (res.author === message.author), {
                                 max: 1,
                                 time: 15000,
-                                errors: ['time'],
+                                errors: ['time']
                             })
-                            .then(() => {
-                                console.log('MAINTENANCE: All commands and aliases evicted!');
-                                message.channel.send(`***Bot Maintenance Mode activated by __${initiator(owner)}__. All commands and aliases deleted from cache.***`)
-                                    .then(() => {
-                                        for (let i = 0; i < cmds.length; i++) {
-                                            if (cmds[i] === 'maintenance')
-                                                continue;
+                            .then(msg => {
 
-                                            if (cmds[i] === 'suggest')
-                                                continue;
+                                let id = msg.first().id;
+                                message.channel.fetchMessage(id).then(async m => {
 
-                                            if (cmds[i] === 'help')
-                                                continue;
+                                    let reason = m.content;
+                                    await message.channel.send(`Your maintenance message: 
+                                    \`\`\`${reason}\`\`\`
+If this is what you want, please type \`confirm\``);
 
-                                            if (cmds[i] === 'beta')
-                                                continue;
+                                    await cmdStatus.set('reason', reason);
+                                });
+                            })
+                            .then(msg => {
 
-                                            client.commands.delete(cmds[i]);
-                                        }
-
-                                        for (let i = 0; i < aliases.length; i++) {
-                                            if (aliases[i] === 'h')
-                                                continue;
-
-                                            if (aliases[i] === 'halp')
-                                                continue;
-
-                                            client.aliases.delete(aliases[i]);
-                                        }
-
-                                        // let path = '../cmdStatus.json';
-                                        // let data = {
-                                        //     "status": 'off'
-                                        // };
-                                        
-                                        // fs.writeFile(path, JSON.stringify(data), (err) => {
-                                        //     if (err) throw err;
-                                        // });
+                                message.channel.awaitMessages(res => res.content === 'confirm', {
+                                        max: 1,
+                                        time: 5000,
+                                        errors: ['time']
                                     })
                                     .then(() => {
-                                        client.user.setActivity(`Maintenance Mode...`, {
-                                                type: 'WATCHING'
+
+                                        console.log('MAINTENANCE: Activated. Locking all non-essential commands...');
+                                        message.channel.send(`***Bot Maintenance Mode activated by __${initiator(owner)}__. All all non-essential commands have been locked.***`)
+                                            .then(() => {
+
+                                                cmdStatus.set('status', 'off');
+
                                             })
-                                            .catch(console.error);
+                                            .then(() => {
+
+                                                client.user.setStatus('online');
+                                                client.user.setActivity('Maintenance Mode...', {
+                                                        type: 'WATCHING'
+                                                    })
+                                                    .catch(console.error);
+                                            })
+                                            .catch(err => console.log(err));
+
                                     })
+                                    .catch(() => {
+                                        message.channel.send('Cancelling. No confirmation...').then(msg => msg.delete(3000));
+                                    });
                             })
-                    })
-                    .catch(() => {
-                        message.channel.send('Failed to confirm within the time period.');
+                            .catch(() => {
+                                message.channel.send('Cancelling. No message...').then(msg => msg.delete(3000));
+                            });
                     });
-                break;
+            break;
             case 'off':
 
+            if (status === 'on') return message.channel.send('Maintenance mode is not currently active.').then(msg => msg.delete(3003));
+
                 message.channel.send('Please type `confirm` if you would like to proceed with ending maintenance mode. You have 15 seconds.')
-                    .then(() => {
-                        message.delete().catch(O_o => {});
-                        message.channel.awaitMessages(res => res.content === 'confirm', {
+                    .then(async () => {
+                        await message.delete().catch(O_o => {});
+                        message.channel.awaitMessages(res => res.content === 'confirm' && res.author === message.author, {
                                 max: 1,
                                 time: 15000,
                                 errors: ['time'],
                             })
                             .then(() => {
-                                console.log('MAINTENANCE: Loading all commands and aliases...');
-                                message.channel.send(`***Bot Maintenance Mode deactivated by __${initiator(owner)}__. All commands and aliases are loaded.***`)
-                                    .then(async () => {
-                                        await fs.readdir('./commands', async (err, files) => {
-                                            if (err) return console.error(err);
-                                            files.forEach(file => {
-                                                if (!file.endsWith('.js')) return;
-                                                let props = require(`./${file}`);
-                                                let cmdName = file.split('.')[0];
-                                                console.log(`Loaded command '${cmdName}'`);
-                                                client.commands.set(cmdName, props);
-                                                props.conf.aliases.forEach(alias => {
-                                                    client.aliases.set(alias, cmdName);
-                                                });
-                                            });
-                                        });
+                                console.log('MAINTENANCE: Deactivated. Unlocking all commands...');
+                                message.channel.send(`***Bot Maintenance Mode deactivated by __${initiator(owner)}__. All commands have been unlocked.***`)
+                                    .then(() => {
 
-                                        // let path = '../cmdStatus.json';
-                                        // let data = {
-                                        //     "status": 'on'
-                                        // };
+                                        cmdStatus.set('status', 'on');
                                         
-                                        // await fs.writeFile(path, JSON.stringify(data), (err) => {
-                                        //     if (err) throw err;
-                                        // });
                                     })
                                     .then(() => {
                                         const userSize = client.users.size.toLocaleString();
@@ -133,24 +120,25 @@ exports.run = async (client, message, args) => {
                                                 type: 'WATCHING'
                                             })
                                             .catch(console.error);
-                                    });
+                                    })
+                                    .catch(err => console.log(err));
                             })
                             .catch(() => {
-                                message.channel.send('Failed to confirm within the time period.');
+                                message.channel.send('Cancelling. No confirmation...').then(msg => msg.delete(3000));
                             });
                     });
                 break;
         }
     });
-}
+};
 
 exports.conf = {
     aliases: [],
-    status: ''
-}
+    status: 'true'
+};
 
 exports.help = {
     name: "maintenance",
     description: "Set the bot into maintenance mode",
     usage: "maintenance <on/off>"
-}
+};
